style: reveal list
