+++ package.json
@@ -38,6 +38,7 @@
     "@metaplex-foundation/umi": "^0.9.1",
     "@metaplex-foundation/umi-bundle-defaults": "^0.9.1",
     "@prisma/client": "^5.10.2",
+    "@radix-ui/react-checkbox": "^1.1.1",
     "@radix-ui/react-slider": "^1.1.2",
     "@solana/spl-token": "^0.4.1",
     "@solana/web3.js": "^1.91.8",
@@ -46,10 +47,13 @@
     "animejs": "^3.2.2",
     "async-mutex": "^0.5.0",
     "bs58": "^5.0.0",
+    "class-variance-authority": "^0.7.0",
     "classnames": "^2.5.1",
+    "clsx": "^2.1.1",
     "date-fns": "^3.6.0",
     "dayjs": "^1.11.10",
     "jsonwebtoken": "^9.0.2",
+    "lucide-react": "^0.417.0",
     "next": "14.1.3",
     "next-nprogress-bar": "^2.3.12",
     "next-pwa": "^5.6.0",
@@ -64,7 +68,8 @@
     "react-spring-bottom-sheet": "^3.4.1",
     "react-virtuoso": "^4.7.7",
     "swiper": "^11.1.3",
-    "tailwind-merge": "^2.3.0",
+    "tailwind-merge": "^2.4.0",
+    "tailwindcss-animate": "^1.0.7",
     "viem": "^2.7.22",
     "zod": "^3.22.4"
   },
